third_party/WebKit/LayoutTests/external/wpt/web-share/idlharness.https-expected.txt

This is a testharness.js-based test.
PASS Test driver 
FAIL Navigator interface: operation share(ShareData) assert_equals: property has wrong .length expected 0 but got 1
PASS Navigator interface: navigator must inherit property "share" with the proper type (0) 
PASS Navigator interface: calling share(ShareData) on navigator with too few arguments must throw TypeError 
Harness: the test ran to completion.
